remove old view
